pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.66.0",
    "nltk==3.9.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.2",
]
